Split out individual unit test into separate test files.
