--- test/langtools/tools/javac/SuperInit/EarlyLocalClass2.java
@@ -1,116 +1,0 @@
-/*
- * Copyright (c) 2024, Oracle and/or its affiliates. All rights reserved.
- * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
- *
- * This code is free software; you can redistribute it and/or modify it
- * under the terms of the GNU General Public License version 2 only, as
- * published by the Free Software Foundation.
- *
- * This code is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
- * version 2 for more details (a copy is included in the LICENSE file that
- * accompanied this code).
- *
- * You should have received a copy of the GNU General Public License version
- * 2 along with this work; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
- *
- * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
- * or visit www.oracle.com if you need additional information or have any
- * questions.
- */
-/*
- * @test
- * @bug 8333313
- * @summary Verify references to local classes declared in early construction contexts
- * @enablePreview
- */
-import java.util.concurrent.atomic.AtomicReference;
-
-public class EarlyLocalClass2 {
-
-    public static class Main1 {
-        class Test {
-            Test() {
-                class InnerLocal { }
-                Runnable r = () -> new InnerLocal();
-                r.run();
-                super();
-            }
-        }
-    }
-
-    public static class Main2 {
-        class Test {
-            Test() {
-                class InnerLocal { }
-                Runnable r = () -> new InnerLocal() { };
-                r.run();
-                super();
-            }
-        }
-    }
-
-    public static class Main3 {
-        class Test {
-            Test() {
-                class InnerLocal { }
-                Runnable r = InnerLocal::new;
-                r.run();
-                super();
-            }
-        }
-    }
-
-    public static class Main4 {
-
-        int y;
-
-        class Test extends AtomicReference<Runnable> {
-            Test(int x) {
-                class Foo implements Runnable {
-                    public void run() {
-                        System.out.println(x + y);
-                    }
-                }
-                super(new Foo());
-            }
-        }
-    }
-
-    public static class Main5 {
-
-        int y;
-
-        class Test extends AtomicReference<Runnable> {
-            Test(int x) {
-                super(new Runnable() {
-                    public void run() {
-                        System.out.println(x + y);
-                    }
-                });
-            }
-        }
-    }
-
-    public static class Main6 {
-
-        int y;
-
-        class Test extends AtomicReference<Runnable> {
-            Test(int x) {
-                super(() -> System.out.println(x + y));
-            }
-        }
-    }
-
-    public static void main(String[] args) {
-        new Main1().new Test();
-        new Main2().new Test();
-        new Main3().new Test();
-        new Main4().new Test(3);
-        new Main5().new Test(3);
-        new Main6().new Test(3);
-    }
-}
